+++ SM_Test.inl
@@ -50,7 +50,7 @@ inline
 bool is_point_in_area(const vec2& v, const std::vector<vec2>& area)
 {
 	bool odd_nodes = false;
-	for (int i = 0, n = area.size(), j = n - 1; i < n; ++i)
+	for (size_t i = 0, n = area.size(), j = n - 1; i < n; ++i)
 	{
 		if (((area[i].y < v.y && area[j].y >= v.y) ||
 			 (area[j].y < v.y && area[i].y >= v.y)) &&
@@ -76,7 +76,7 @@ bool is_point_in_convex(const vec2& pos, const std::vector<vec2>& convex)
 }
 
 inline
-bool is_point_in_convex(const vec2& pos, const vec2* convex, int num)
+bool is_point_in_convex(const vec2& pos, const vec2* convex, size_t num)
 {
 	if (num < 3) {
 		return false;
